Move remaining cms and www routes.
