+++ politeiawww/legacy/codestats_test.go
@@ -1,4 +1,4 @@
-package main
+package legacy
 
 import (
 	"fmt"
