+++ politeiawww/legacy/www.go
@@ -1,0 +1,109 @@
+// Copyright (c) 2017-2020 The Decred developers
+// Use of this source code is governed by an ISC
+// license that can be found in the LICENSE file.
+
+package legacy
+
+import (
+	"encoding/hex"
+	"encoding/json"
+	"fmt"
+	"net/http"
+	"net/http/httputil"
+
+	"github.com/decred/politeia/politeiad/api/v1/mime"
+	"github.com/decred/politeia/util"
+	"github.com/decred/politeia/util/version"
+	"github.com/gorilla/csrf"
+)
+
+// handleNotFound is a generic handler for an invalid route.
+func (p *LegacyPoliteiawww) handleNotFound(w http.ResponseWriter, r *http.Request) {
+	// Log incoming connection
+	log.Debugf("Invalid route: %v %v %v %v",
+		util.RemoteAddr(r), r.Method, r.URL, r.Proto)
+
+	// Trace incoming request
+	log.Tracef("%v", newLogClosure(func() string {
+		trace, err := httputil.DumpRequest(r, true)
+		if err != nil {
+			trace = []byte(fmt.Sprintf("logging: "+
+				"DumpRequest %v", err))
+		}
+		return string(trace)
+	}))
+
+	util.RespondWithJSON(w, http.StatusNotFound, www.ErrorReply{})
+}
+
+// version is an HTTP GET to determine the lowest API route version that this
+// backend supports.  Additionally it is used to obtain a CSRF token.
+func (p *LegacyPoliteiawww) handleVersion(w http.ResponseWriter, r *http.Request) {
+	log.Tracef("handleVersion")
+
+	versionReply := www.VersionReply{
+		Version:      www.PoliteiaWWWAPIVersion,
+		Route:        www.PoliteiaWWWAPIRoute,
+		BuildVersion: version.BuildMainVersion(),
+		PubKey:       hex.EncodeToString(p.cfg.Identity.Key[:]),
+		TestNet:      p.cfg.TestNet,
+		Mode:         p.cfg.Mode,
+	}
+
+	_, err := p.sessions.GetSessionUser(w, r)
+	if err == nil {
+		versionReply.ActiveUserSession = true
+	}
+
+	vr, err := json.Marshal(versionReply)
+	if err != nil {
+		RespondWithError(w, r, 0, "handleVersion: Marshal %v", err)
+		return
+	}
+
+	w.Header().Set("Strict-Transport-Security",
+		"max-age=63072000; includeSubDomains")
+	w.Header().Set("X-Content-Type-Options", "nosniff")
+	w.Header().Set("Referrer-Policy", "same-origin")
+	w.Header().Set("X-Frame-Options", "DENY")
+	w.Header().Set("X-XSS-Protection", "1; mode=block")
+	w.Header().Set("Content-Type", "application/json; charset=utf-8")
+	w.Header().Set(www.CsrfToken, csrf.Token(r))
+
+	w.WriteHeader(http.StatusOK)
+	w.Write(vr)
+}
+
+func (p *LegacyPoliteiawww) handlePolicy(w http.ResponseWriter, r *http.Request) {
+	// Get the policy command.
+	log.Tracef("handlePolicy")
+
+	reply := &www.PolicyReply{
+		MinPasswordLength:          www.PolicyMinPasswordLength,
+		MinUsernameLength:          www.PolicyMinUsernameLength,
+		MaxUsernameLength:          www.PolicyMaxUsernameLength,
+		UsernameSupportedChars:     www.PolicyUsernameSupportedChars,
+		ProposalListPageSize:       www.ProposalListPageSize,
+		UserListPageSize:           www.UserListPageSize,
+		MaxImages:                  www.PolicyMaxImages,
+		MaxImageSize:               www.PolicyMaxImageSize,
+		MaxMDs:                     www.PolicyMaxMDs,
+		MaxMDSize:                  www.PolicyMaxMDSize,
+		PaywallEnabled:             p.paywallIsEnabled(),
+		ValidMIMETypes:             mime.ValidMimeTypes(),
+		MinProposalNameLength:      www.PolicyMinProposalNameLength,
+		MaxProposalNameLength:      www.PolicyMaxProposalNameLength,
+		ProposalNameSupportedChars: www.PolicyProposalNameSupportedChars,
+		MaxCommentLength:           www.PolicyMaxCommentLength,
+		TokenPrefixLength:          www.TokenPrefixLength,
+		BuildInformation:           version.BuildInformation(),
+		IndexFilename:              www.PolicyIndexFilename,
+		MinLinkByPeriod:            0,
+		MaxLinkByPeriod:            0,
+		MinVoteDuration:            0,
+		MaxVoteDuration:            0,
+		PaywallConfirmations:       p.cfg.MinConfirmationsRequired,
+	}
+
+	util.RespondWithJSON(w, http.StatusOK, reply)
+}
