+++ politeiawww/legacy/www_test.go
@@ -1,4 +1,8 @@
-package main
+// Copyright (c) 2017-2020 The Decred developers
+// Use of this source code is governed by an ISC
+// license that can be found in the LICENSE file.
+
+package legacy
 
 import (
 	"encoding/hex"
